deepsearch/backend/requirements.txt
flask
crewai
crewai_tools
python-dotenv 